fix: move error logic to handleCloudFormationError (add tests)
